Refactor global (#1808)
