+++ internal/controller/datadogagent/controller_reconcile_v2.go
@@ -66,6 +66,9 @@ func (r *Reconciler) reconcileInstanceV2(ctx context.Context, logger logr.Logger
 	depsStore, resourceManagers := r.setupDependencies(instance, logger)
 
 	var err error
+	if err = r.manageGlobalDependencies(logger, instance, resourceManagers, requiredComponents); err != nil {
+		return r.updateStatusIfNeededV2(logger, instance, newStatus, reconcile.Result{}, err, now)
+	}
 	if err = r.manageFeatureDependencies(logger, enabledFeatures, requiredComponents, resourceManagers); err != nil {
 		return r.updateStatusIfNeededV2(logger, instance, newStatus, reconcile.Result{}, err, now)
 	}
