+++ internal/controller/datadogagent/global/const.go
@@ -1,0 +1,17 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2016-present Datadog, Inc.
+
+package global
+
+// This file tracks constants used in global
+
+const (
+	kubeletCAVolumeName = "kubelet-ca"
+
+	FIPSProxyCustomConfigVolumeName = "fips-proxy-cfg"
+	FIPSProxyCustomConfigFileName   = "datadog-fips-proxy.cfg"
+	FIPSProxyCustomConfigMapName    = "%s-fips-config"
+	FIPSProxyCustomConfigMountPath  = "/etc/datadog-fips-proxy/datadog-fips-proxy.cfg"
+)
