+++ internal/controller/datadogagent/override/podtemplatespec_test.go
@@ -20,6 +20,7 @@ import (
 	apiutils "github.com/DataDog/datadog-operator/api/utils"
 	"github.com/DataDog/datadog-operator/internal/controller/datadogagent/common"
 	"github.com/DataDog/datadog-operator/internal/controller/datadogagent/feature/fake"
+	"github.com/DataDog/datadog-operator/pkg/constants"
 )
 
 func TestPodTemplateSpec(t *testing.T) {
@@ -589,7 +590,7 @@ func TestPodTemplateSpec(t *testing.T) {
 				manager.EnvVarMgr.AddEnvVarToContainer(
 					apicommon.ClusterAgentContainerName,
 					&v1.EnvVar{
-						Name:  DDLogLevel,
+						Name:  constants.DDLogLevel,
 						Value: "info",
 					},
 				)
@@ -607,7 +608,7 @@ func TestPodTemplateSpec(t *testing.T) {
 				envSet := false
 
 				for _, env := range manager.EnvVarMgr.EnvVarsByC[apicommon.ClusterAgentContainerName] {
-					if env.Name == DDLogLevel && env.Value == "trace" {
+					if env.Name == constants.DDLogLevel && env.Value == "trace" {
 						envSet = true
 						break
 					}
