+++ internal/controller/datadogagent/override/utils.go
@@ -10,8 +10,6 @@ import (
 	"strings"
 
 	corev1 "k8s.io/api/core/v1"
-
-	"github.com/DataDog/datadog-operator/internal/controller/datadogagent/object"
 )
 
 func getDefaultConfigMapName(ddaName, fileName string) string {
@@ -24,8 +22,4 @@ func hasProbeHandler(probe *corev1.Probe) bool {
 		return true
 	}
 	return false
-}
-
-func getDCATokenChecksumAnnotationKey() string {
-	return object.GetChecksumAnnotationKey("dca-token")
 }
